Change id of `Yes` radio
When forms are in an error state, the Error Summary component will
contain links to the field which uses the field's id.  For radios,
it should target the first option (`Yes` in the case of Yes/No
radios).  Setting the id of the first option to equal the name of
the field makes producing this link correctly much easier.